Add configurable httpd timeout
This patch adds an apiTimeout field to the DesignateSpecCore to
DesignateSpec, to allow a configuration of the timeouts for HAProxy and
Apache.
Inherit from the controller in case it is not specified by the user.

Jira: https://issues.redhat.com/browse/OSPRH-10963
