+++ templates/designateapi/config/httpd.conf
@@ -23,6 +23,8 @@ SetEnvIf X-Forwarded-For "^.*\..*\..*\..*" forwarded
 CustomLog /dev/stdout combined env=!forwarded
 CustomLog /dev/stdout proxy env=forwarded
 
+TimeOut {{ $.TimeOut }}
+
 {{ range $endpt, $vhost := .VHosts }}
   # {{ $endpt }} vhost {{ $vhost.ServerName }} configuration
   <VirtualHost *:9001>
